Removed unnecessary conditional methods from Surveys screen
